removed debug tag for animated scale
