--- lib/features/home/home_screen.dart
@@ -1,144 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
-import 'package:productive/common/resources/strings.dart';
-import 'package:productive/utils/extensions/gradient_extensions.dart';
-
-class HomeScreen extends StatelessWidget {
-  const HomeScreen();
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: const Center(
-        child: Text("Hello world"),
-      ),
-      appBar: AppBar(
-          title: const AppbarContent(),
-          automaticallyImplyLeading: false),
-      floatingActionButton: const Fab(),
-      floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
-      bottomNavigationBar: const BottomNav(),
-    );
-  }
-}
-
-class Fab extends StatelessWidget {
-  const Fab({
-    Key key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 6),
-        child: RaisedButton(
-          onPressed: () {},
-          shape:
-              RoundedRectangleBorder(borderRadius: BorderRadius.circular(80.0)),
-          padding: const EdgeInsets.all(0.0),
-          child: const Padding(
-            padding: EdgeInsets.symmetric(horizontal: 16.0, vertical: 8),
-            child: Text(
-              '+ Add Todo',
-              textAlign: TextAlign.center,
-              style: TextStyle(
-                  fontSize: 18,
-                  color: Colors.white,
-                  fontWeight: FontWeight.w600),
-            ),
-          ).withBlueGradientBg(),
-        ));
-  }
-}
-
-class BottomNav extends StatefulWidget {
-  const BottomNav({
-    Key key,
-  }) : super(key: key);
-
-  @override
-  _BottomNavState createState() => _BottomNavState();
-}
-
-class _BottomNavState extends State<BottomNav> {
-  int selectedIndex = 0;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      decoration: const BoxDecoration(boxShadow: [
-        BoxShadow(
-            color: Color(0x33CCCCCC),
-            offset: Offset(-4, -4),
-            blurRadius: 8,
-            spreadRadius: 8)
-      ]),
-      child: ClipRRect(
-        borderRadius: const BorderRadius.only(
-          topLeft: Radius.circular(32),
-          topRight: Radius.circular(32),
-        ),
-        child: BottomAppBar(
-          elevation: 20,
-          shape: const AutomaticNotchedShape(
-            RoundedRectangleBorder(),
-            StadiumBorder(
-              side: BorderSide(),
-            ),
-          ),
-          child: Container(
-            padding: const EdgeInsets.all(16.0),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: <Widget>[
-                Padding(
-                  padding: const EdgeInsets.symmetric(horizontal: 16.0),
-                  child: SvgPicture.asset(
-                    R.svgImages.homeIcon,
-                  ).withGreyGradientFg(),
-                ),
-                Padding(
-                  padding: const EdgeInsets.symmetric(horizontal: 16.0),
-                  child: SvgPicture.asset(
-                    R.svgImages.statsIcon,
-                  ).withBlueGradientFg(),
-                )
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class AppbarContent extends StatelessWidget {
-  const AppbarContent({
-    Key key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: <Widget>[
-        ClipRRect(
-          borderRadius: BorderRadius.circular(200),
-          child: Image.network(johnDoeImage, height: 52, width: 52),
-        ),
-        const Text(
-          "Todo",
-          style: TextStyle(fontSize: 28, fontWeight: FontWeight.w600),
-        ),
-        SvgPicture.asset(
-          "assets/images/success.svg",
-          height: 32,
-          width: 32,
-        )
-      ],
-    );
-  }
-}
-
-String johnDoeImage =
-    "https://themes.themewaves.com/nuzi/wp-content/uploads/sites/4/2013/05/Team-Member-3.jpg";
