+++ lib/features/home/widgets/home_body.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class HomeBody extends StatelessWidget {
+  const HomeBody({Key key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text("Hello world"),
+    );
+  }
+}
