+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:productive/features/home/home_screen.dart';
+import 'package:productive/features/home/widgets/home_screen.dart';
 import 'package:productive/features/splash/splash_screen.dart';
 import 'package:productive/utils/design/colors.dart';
 
@@ -14,7 +14,7 @@ class ProductiveApp extends StatelessWidget {
       theme: ThemeData(
           fontFamily: 'Product Sans', primarySwatch: whiteMaterialColor),
       debugShowCheckedModeBanner: false,
-      home: const SplashScreen(),
+      home: const HomeScreen(),
     );
   }
 }
